+++ xchange-ftx/src/main/java/org/knowm/xchange/ftx/service/FtxTradeService.java
@@ -1,27 +1,19 @@
 package org.knowm.xchange.ftx.service;
 
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
 import org.knowm.xchange.Exchange;
-import org.knowm.xchange.currency.CurrencyPair;
 import org.knowm.xchange.dto.Order;
 import org.knowm.xchange.dto.account.OpenPositions;
 import org.knowm.xchange.dto.trade.LimitOrder;
 import org.knowm.xchange.dto.trade.MarketOrder;
 import org.knowm.xchange.dto.trade.OpenOrders;
 import org.knowm.xchange.dto.trade.UserTrades;
-import org.knowm.xchange.ftx.FtxAdapters;
-import org.knowm.xchange.ftx.dto.trade.CancelAllFtxOrdersParams;
 import org.knowm.xchange.service.trade.TradeService;
-import org.knowm.xchange.service.trade.params.CancelOrderByCurrencyPair;
 import org.knowm.xchange.service.trade.params.CancelOrderParams;
-import org.knowm.xchange.service.trade.params.CurrencyPairParam;
-import org.knowm.xchange.service.trade.params.TradeHistoryParamCurrencyPair;
-import org.knowm.xchange.service.trade.params.TradeHistoryParamInstrument;
 import org.knowm.xchange.service.trade.params.TradeHistoryParams;
 import org.knowm.xchange.service.trade.params.orders.OpenOrdersParams;
+
+import java.io.IOException;
+import java.util.Collection;
 
 public class FtxTradeService extends FtxTradeServiceRaw implements TradeService {
 
@@ -31,100 +23,46 @@ public class FtxTradeService extends FtxTradeServiceRaw implements TradeService 
 
   @Override
   public String placeMarketOrder(MarketOrder marketOrder) throws IOException {
-    return placeMarketOrderForSubaccount(null, marketOrder);
-  }
-
-  public String placeMarketOrderForSubaccount(String subaccount, MarketOrder marketOrder)
-      throws IOException {
-    return placeNewFtxOrder(subaccount, FtxAdapters.adaptMarketOrderToFtxOrderPayload(marketOrder))
-        .getResult()
-        .getId();
+    return placeMarketOrderForSubaccount(exchange.getExchangeSpecification().getUserName(), marketOrder);
   }
 
   @Override
   public String placeLimitOrder(LimitOrder limitOrder) throws IOException {
-    return placeLimitOrderForSubaccount(null, limitOrder);
-  }
-
-  public String placeLimitOrderForSubaccount(String subaccount, LimitOrder limitOrder)
-      throws IOException {
-    return placeNewFtxOrder(subaccount, FtxAdapters.adaptLimitOrderToFtxOrderPayload(limitOrder))
-        .getResult()
-        .getId();
+    return placeLimitOrderForSubaccount(exchange.getExchangeSpecification().getUserName(), limitOrder);
   }
 
   @Override
   public UserTrades getTradeHistory(TradeHistoryParams params) throws IOException {
-
-    if (params instanceof TradeHistoryParamCurrencyPair) {
-      return FtxAdapters.adaptUserTrades(
-          getFtxOrderHistory(
-                  FtxAdapters.adaptCurrencyPairToFtxMarket(
-                      ((TradeHistoryParamCurrencyPair) params).getCurrencyPair()))
-              .getResult());
-    } else if (params instanceof TradeHistoryParamInstrument) {
-      CurrencyPair currencyPair =
-          new CurrencyPair(((TradeHistoryParamInstrument) params).getInstrument().toString());
-      return FtxAdapters.adaptUserTrades(
-          getFtxOrderHistory(FtxAdapters.adaptCurrencyPairToFtxMarket(currencyPair)).getResult());
-    } else {
-      throw new IOException(
-          "TradeHistoryParams must implement TradeHistoryParamCurrencyPair or TradeHistoryParamInstrument interface.");
-    }
+    return getTradeHistoryForSubaccount(exchange.getExchangeSpecification().getUserName(), params);
   }
 
   @Override
   public boolean cancelOrder(String orderId) throws IOException {
-    return cancelFtxOrder(orderId);
+    return cancelOrderForSubaccount(exchange.getExchangeSpecification().getUserName(), orderId);
   }
 
   @Override
   public boolean cancelOrder(CancelOrderParams orderParams) throws IOException {
-    if (orderParams instanceof CancelOrderByCurrencyPair) {
-      return cancelAllFtxOrders(
-          new CancelAllFtxOrdersParams(
-              FtxAdapters.adaptCurrencyPairToFtxMarket(
-                  ((CancelOrderByCurrencyPair) orderParams).getCurrencyPair())));
-    } else {
-      throw new IOException(
-          "CancelOrderParams must implement CancelOrderByCurrencyPair interface.");
-    }
+    return cancelOrderForSubaccount(exchange.getExchangeSpecification().getUserName(), orderParams);
   }
 
   @Override
   public Collection<Order> getOrder(String... orderIds) throws IOException {
-    return getOrderFromSubaccount(null, orderIds);
-  }
-
-  public Collection<Order> getOrderFromSubaccount(String subaccount, String... orderIds)
-      throws IOException {
-    List<Order> orderList = new ArrayList<>();
-    for (String orderId : orderIds) {
-      Order order = FtxAdapters.adaptLimitOrder(getFtxOrderStatus(subaccount, orderId).getResult());
-      orderList.add(order);
-    }
-    return orderList;
+    return getOrderFromSubaccount(exchange.getExchangeSpecification().getUserName(), orderIds);
   }
 
   @Override
   public OpenOrders getOpenOrders(OpenOrdersParams params) throws IOException {
-    if (params instanceof CurrencyPairParam) {
-      return FtxAdapters.adaptOpenOrders(
-          getFtxOpenOrders(
-              FtxAdapters.adaptCurrencyPairToFtxMarket(
-                  ((CurrencyPairParam) params).getCurrencyPair())));
-    } else {
-      throw new IOException("OpenOrdersParams must implement CurrencyPairParam interface.");
-    }
+    return getOpenOrdersForSubaccount(exchange.getExchangeSpecification().getUserName(), params);
   }
 
   @Override
   public OpenOrders getOpenOrders() throws IOException {
-    return FtxAdapters.adaptOpenOrders(getFtxAllOpenOrders());
+    return getOpenOrdersForSubaccount(exchange.getExchangeSpecification().getUserName());
   }
 
   @Override
   public OpenPositions getOpenPositions() throws IOException {
-    return FtxAdapters.adaptOpenPositions(getFtxPositions().getResult());
+    return getOpenPositionsForSubaccount(exchange.getExchangeSpecification().getUserName());
   }
 }
